fix(批次266): 修复 paginate page_size 类型错误（usize→u64）
SeaORM 1.1.20 的 .paginate() 方法签名要求 page_size: u64，
原修复误用 as usize 导致 E0308 类型不匹配编译失败。

+++ backend/src/services/fixed_asset_service.rs
@@ -138,7 +138,7 @@ impl FixedAssetService {
         // 补 page_size.clamp(1, 100) 防 DoS（原实现仅 clamp page，page_size 无上限保护）
         let paginator = query
             .order_by(fixed_asset::Column::Id, Order::Desc)
-            .paginate(&*self.db, params.page_size.clamp(1, 100) as usize);
+            .paginate(&*self.db, params.page_size.clamp(1, 100) as u64);
         let (assets, total) =
             paginate_with_total(paginator, params.page.clamp(1, 1000) as u64).await?;
